backend/pyproject.toml
[project]
name = "neuromarketing"
version = "0.1.0"
description = "AI Neuro-Marketing Validator - predict brain responses to marketing videos"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.0.0",
    "anthropic>=0.40.0",
    "ffmpeg-python>=0.2.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
]
gpu = [
    "torch>=2.5.0,<2.7.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "gpu: requires GPU for TRIBE v2 inference",
    "live_api: hits real external APIs (Claude, HuggingFace)",
]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
]
